hoist template migrations
This moves the codemods for the templates up 1 directory. This will make
it consistent with the CSS and Config codemods setup.

This now means that `./src/codemods/template/codemods/*` becomes `./src/codemods/template/*`.
