+++ packages/@tailwindcss-upgrade/src/codemods/css/migrate-theme-to-var.ts
@@ -1,6 +1,6 @@
 import { type Plugin } from 'postcss'
 import type { DesignSystem } from '../../../../tailwindcss/src/design-system'
-import { Convert, createConverter } from '../template/codemods/migrate-theme-to-var'
+import { Convert, createConverter } from '../template/migrate-theme-to-var'
 
 export function migrateThemeToVar({
   designSystem,
