+++ packages/@tailwindcss-upgrade/src/codemods/template/migrate-automatic-var-injection.test.ts
