+++ packages/@tailwindcss-upgrade/src/codemods/template/migrate-automatic-var-injection.ts
@@ -1,8 +1,8 @@
-import { walk, WalkAction } from '../../../../../tailwindcss/src/ast'
-import { type Candidate, type Variant } from '../../../../../tailwindcss/src/candidate'
-import type { Config } from '../../../../../tailwindcss/src/compat/plugin-api'
-import type { DesignSystem } from '../../../../../tailwindcss/src/design-system'
-import { printCandidate } from '../candidates'
+import { walk, WalkAction } from '../../../../tailwindcss/src/ast'
+import { type Candidate, type Variant } from '../../../../tailwindcss/src/candidate'
+import type { Config } from '../../../../tailwindcss/src/compat/plugin-api'
+import type { DesignSystem } from '../../../../tailwindcss/src/design-system'
+import { printCandidate } from './candidates'
 
 export function automaticVarInjection(
   designSystem: DesignSystem,
