+++ packages/@tailwindcss-upgrade/src/codemods/template/migrate-legacy-arbitrary-values.ts
@@ -1,8 +1,8 @@
-import { parseCandidate } from '../../../../../tailwindcss/src/candidate'
-import type { Config } from '../../../../../tailwindcss/src/compat/plugin-api'
-import type { DesignSystem } from '../../../../../tailwindcss/src/design-system'
-import { segment } from '../../../../../tailwindcss/src/utils/segment'
-import { printCandidate } from '../candidates'
+import { parseCandidate } from '../../../../tailwindcss/src/candidate'
+import type { Config } from '../../../../tailwindcss/src/compat/plugin-api'
+import type { DesignSystem } from '../../../../tailwindcss/src/design-system'
+import { segment } from '../../../../tailwindcss/src/utils/segment'
+import { printCandidate } from './candidates'
 
 export function legacyArbitraryValues(
   designSystem: DesignSystem,
