+++ packages/@tailwindcss-upgrade/src/codemods/template/migrate-max-width-screen.ts
@@ -1,6 +1,6 @@
-import type { Config } from '../../../../../tailwindcss/src/compat/plugin-api'
-import type { DesignSystem } from '../../../../../tailwindcss/src/design-system'
-import { printCandidate } from '../candidates'
+import type { Config } from '../../../../tailwindcss/src/compat/plugin-api'
+import type { DesignSystem } from '../../../../tailwindcss/src/design-system'
+import { printCandidate } from './candidates'
 
 export function maxWidthScreen(
   designSystem: DesignSystem,
