+++ packages/@tailwindcss-upgrade/src/codemods/template/migrate-prefix.ts
@@ -1,8 +1,8 @@
-import { parseCandidate, type Candidate } from '../../../../../tailwindcss/src/candidate'
-import type { Config } from '../../../../../tailwindcss/src/compat/plugin-api'
-import type { DesignSystem } from '../../../../../tailwindcss/src/design-system'
-import { segment } from '../../../../../tailwindcss/src/utils/segment'
-import { printCandidate } from '../candidates'
+import { parseCandidate, type Candidate } from '../../../../tailwindcss/src/candidate'
+import type { Config } from '../../../../tailwindcss/src/compat/plugin-api'
+import type { DesignSystem } from '../../../../tailwindcss/src/design-system'
+import { segment } from '../../../../tailwindcss/src/utils/segment'
+import { printCandidate } from './candidates'
 
 let seenDesignSystems = new WeakSet<DesignSystem>()
