+++ packages/@tailwindcss-upgrade/src/codemods/template/migrate-simple-legacy-classes.test.ts
