+++ packages/@tailwindcss-upgrade/src/codemods/template/migrate-theme-to-var.test.ts
