+++ packages/@tailwindcss-upgrade/src/codemods/template/migrate-variant-order.ts
@@ -1,8 +1,8 @@
-import { walk, type AstNode } from '../../../../../tailwindcss/src/ast'
-import { type Variant } from '../../../../../tailwindcss/src/candidate'
-import type { Config } from '../../../../../tailwindcss/src/compat/plugin-api'
-import type { DesignSystem } from '../../../../../tailwindcss/src/design-system'
-import { printCandidate } from '../candidates'
+import { walk, type AstNode } from '../../../../tailwindcss/src/ast'
+import { type Variant } from '../../../../tailwindcss/src/candidate'
+import type { Config } from '../../../../tailwindcss/src/compat/plugin-api'
+import type { DesignSystem } from '../../../../tailwindcss/src/design-system'
+import { printCandidate } from './candidates'
 
 export function variantOrder(
   designSystem: DesignSystem,
