+++ packages/@tailwindcss-upgrade/src/codemods/template/migrate.ts
@@ -4,19 +4,19 @@ import type { Config } from '../../../../tailwindcss/src/compat/plugin-api'
 import type { DesignSystem } from '../../../../tailwindcss/src/design-system'
 import { spliceChangesIntoString, type StringChange } from '../../utils/splice-changes-into-string'
 import { extractRawCandidates } from './candidates'
-import { arbitraryValueToBareValue } from './codemods/migrate-arbitrary-value-to-bare-value'
-import { automaticVarInjection } from './codemods/migrate-automatic-var-injection'
-import { bgGradient } from './codemods/migrate-bg-gradient'
-import { handleEmptyArbitraryValues } from './codemods/migrate-handle-empty-arbitrary-values'
-import { important } from './codemods/migrate-important'
-import { legacyArbitraryValues } from './codemods/migrate-legacy-arbitrary-values'
-import { legacyClasses } from './codemods/migrate-legacy-classes'
-import { maxWidthScreen } from './codemods/migrate-max-width-screen'
-import { modernizeArbitraryValues } from './codemods/migrate-modernize-arbitrary-values'
-import { prefix } from './codemods/migrate-prefix'
-import { simpleLegacyClasses } from './codemods/migrate-simple-legacy-classes'
-import { themeToVar } from './codemods/migrate-theme-to-var'
-import { variantOrder } from './codemods/migrate-variant-order'
+import { arbitraryValueToBareValue } from './migrate-arbitrary-value-to-bare-value'
+import { automaticVarInjection } from './migrate-automatic-var-injection'
+import { bgGradient } from './migrate-bg-gradient'
+import { handleEmptyArbitraryValues } from './migrate-handle-empty-arbitrary-values'
+import { important } from './migrate-important'
+import { legacyArbitraryValues } from './migrate-legacy-arbitrary-values'
+import { legacyClasses } from './migrate-legacy-classes'
+import { maxWidthScreen } from './migrate-max-width-screen'
+import { modernizeArbitraryValues } from './migrate-modernize-arbitrary-values'
+import { prefix } from './migrate-prefix'
+import { simpleLegacyClasses } from './migrate-simple-legacy-classes'
+import { themeToVar } from './migrate-theme-to-var'
+import { variantOrder } from './migrate-variant-order'
 
 export type Migration = (
   designSystem: DesignSystem,
